requirements.txt
fonttools
python-dotenv
pyserial